refactored service classes
